Organize the project according to best practice
Organize the project structure according to best practices.

* **Add new modules**:
  - Add `http_client.py` with `HTTPClient` and `IHTTPClient` classes.
  - Add `form_handler.py` with `FormHandler` class.
  - Add `cookie_handler.py` with `CookieHandler` class.

* **Move configuration files**:
  - Move `logger_config.py` to `config/logger_config.py`.
  - Move `config.json` to `config/config.json`.
  - Move `postdata.json` to `config/postdata.json`.

* **Initialize config package**:
  - Add `config/__init__.py` to initialize the package and load configuration files.

* **Update imports and paths in main.py**:
  - Import `HTTPClient`, `FormHandler`, and `CookieHandler` from their respective new modules.
  - Update paths for `config` imports.
